speedup table for 29 qubit 3sat on cx2

+++ 3sat/arcus-b_CPU/35qubits_32rankx16threads/3SATGenerator.h
@@ -20,7 +20,7 @@ int getNumClauses(int numBools);
 
 void nextBinaryNumber(int *bits, int numBits);
 
-void convertToBinary(int number, int numBits, int *outputBits);
+void convertToBinary(long long int number, int numBits, int *outputBits);
 
 
 #endif // SAT_GEN_H_
